+++ source/common/quic/quic_filter_manager_connection_impl.cc
@@ -11,7 +11,7 @@ namespace Quic {
 
 QuicFilterManagerConnectionImpl::QuicFilterManagerConnectionImpl(
     QuicNetworkConnection& connection, const quic::QuicConnectionId& connection_id,
-    Event::Dispatcher& dispatcher, uint32_t send_buffer_limit,
+    Event::Dispatcher& dispatcher, uint64_t send_buffer_limit,
     std::shared_ptr<QuicSslConnectionInfo>&& info)
     // Using this for purpose other than logging is not safe. Because QUIC connection id can be
     // 18 bytes, so there might be collision when it's hashed to 8 bytes.
@@ -131,10 +131,10 @@ void QuicFilterManagerConnectionImpl::rawWrite(Buffer::Instance& /*data*/, bool 
   IS_ENVOY_BUG("unexpected call to rawWrite");
 }
 
-void QuicFilterManagerConnectionImpl::updateBytesBuffered(size_t old_buffered_bytes,
-                                                          size_t new_buffered_bytes) {
+void QuicFilterManagerConnectionImpl::updateBytesBuffered(uint64_t old_buffered_bytes,
+                                                          uint64_t new_buffered_bytes) {
   int64_t delta = new_buffered_bytes - old_buffered_bytes;
-  const size_t bytes_to_send_old = bytes_to_send_;
+  const uint64_t bytes_to_send_old = bytes_to_send_;
   bytes_to_send_ += delta;
   if (delta < 0) {
     ENVOY_BUG(bytes_to_send_old > bytes_to_send_, "Underflowed");
